[Xamarin.Android.Build.Tasks] Add metadata to library imports
One of the hard things to do when diagnosing issues
is figuring out where a certain file came from.
Because of the path and command line limits on
Windows we have to use shortened paths by default.
So we have paths such as

	obj/Debug/lp/32/res
	obj/Debug/lp/42/classes.jar
	obj/Debug/lp/55/assets

While this is great for path lengths, it does not
tell use which nuget or Library project those
items came from. There is a map file which does
hold the mapping information, however it is not
a straight forward process to look at that file and
then work out the path to the assembly or `.aar`.

So what this PR does is add a new piece of metadata
to each item specifying the `OriginalFile` which
the item came from. This will either be a full path
to an Assembly, for items which are extracted from
an EmbeddedResource. Or a full path to an `.aar` file,
which was downloaded with a nuget.

With this information diagnosing issues might get a
bit easier.