[Xamarin.Android.Build.Tests] Migrate Tests from commercial repo (#2980)
As part of a move to unify our unit tests into the main repo this PR
brings in all the InstantRun tests.  They are protected behind a
`CommercialBuildAvailable` check which will only be `true` if the
commercial parts are available. The commercial tests are also flagged
with a nunit `[Category("Commercial")]` just in case we decide to
filter in the future.

This PR also makes some changes to the folder structure.  ALL `Task`
based tests (i.e ones that test just a `Task` using the
`MockBbuildEngine`) have been moved into the `Tasks` folder.
We already have a few tests in that folder, this make sure they are
all in one place.

The `*.OSS.cs` partial class files have been renamed to
`*.TestCaseSource.cs`.  They also handle the tests for BOTH open
source and commercial builds.  As a result they will not longer be
needed in the commercial repo.  They were also moved into the shared
project.  This change will probably cause commercial build failures
until PR #966 is merged on the commercial side.