[tests] Split APK tests into parallel and non-parallel runs
We want to run 2 apk tests, locale and XF, non-parallel. So that the
performance measurements for these tests make more sense and are
comparable between test flavors and also between builds.

The splitting is implemented by introducing new
`RunPerformanceApkTests` target and pre-filtering `TestApk`,
`TestApkInstrumentation` and `TestApkPermission` item groups for this
target and also for `RunApkTests` target.