[Xamarin.Android.Build.Tasks] fix arbitrary files in App Bundles (#2989)
Fixes: https://github.com/xamarin/xamarin-android/issues/2969

If you tried to use our new Android App Bundle feature with an app
using Google Play Services, you would get an error such as:

    [BT Module files can be only in pre-defined directories, but found 'protobuf.meta'.

Looking at `obj\Release\android\bin\base.zip`, it indeed had a file
named `protobuf.meta` in the root of the zip file.

In 4f08867, I missed a case where `<BuildApk/>` puts arbitrary files
from `.jar` files into `.apk` or `.aab` archives. For this to work
with app bundles, files like this need to go into a `/root/`
directory. I have a `RootPath` property to handle this, which I forgot
to use in this case.

~~ Tests ~~

I updated `BundleToolTests` so that we use a
`XamarinFormsMapsApplicationProject` that contains Google Play
Services.

This gave us a couple benefits:

* We now have test coverage of these arbitrary files.
* We now have a new "split" based on language: base-en.apk

I also updated the tests so that *new* files won't cause the tests to
fail.

~~ Future changes ~~

After seeing this code in `<BuildApk/>`, I want to improve it. It uses
`byte[]` + LINQ, I think there are some perf improvements to be done
here.