[tests] check d8 works with "Enhanced Fast Deployment"
There is a direct call to `<CompileToDalvik/>` when using
`AndroidFastDeploymentType` = `Assemblies:Dexes`.

We need to add a few tests here to verify that this code path still
works when using `$(AndroidDexTool)` = `d8`.

I also did some minor cleanup in these tests, such as:

* Instead of checking if the `classes.dex` < 30K, I actually checked
  for a class that should be missing and one that should exist.
* Use the `TestName` property.
* Use of `CreateApkBuilder (..., false, false)` can just omit the two
  `false, false` parameters, as that is the default now.

The actual changes for d8 + enhanced fastdev will be in monodroid.