[mono-runtimes] Fix the timestamps on the mono bundle.
We had a weird bunch of test failures on one
of the PR's which bump mono. Many of the
MSBuild tests were failing becuase targets
were running when they shouldn't. Looking
at the logs we saw

	Input file "/Users/builder/jenkins/workspace/xamarin-android-pr-pipeline-release/xamarin-android/bin/Release/lib/xamarin.android/xbuild-frameworks/MonoAndroid/v1.0/Facades/Microsoft.Win32.Primitives.dll" is newer than output file "obj/Debug/UnnamedProject.dll".

When ment the `Microsoft.Win32.Primitives.dll` assembly
somehow was `newer` than a file we just build...
Further investigation into the mono bundle resulted in

	 unzip -l android-release-Darwin-cf880be66a2d1cc8ca34e345114a7f420b7b86ea.zip | grep Microsoft.Win32
		4608  04-17-2019 08:54   android-bcl/monodroid/Facades/Microsoft.Win32.Primitives.dll

So this file was created on `04-17-2019 08:54`. However the
unit test build log had

	Build started 4/16/2019 7:50:57 PM.

So the mono bundle had different timestamps. This is
probably down to something weird on the bots.
So in order to try to fix this we shall `Touch` the
files we get from the bundle to make sure they have
the current date.