[mono-runtimes] Fix timestamps from the mono archive (#3003)
Context: https://jenkins.mono-project.com/job/xamarin-android-pr-pipeline-release/632/

We had a weird bunch of test failures on PR #2701, which bumps to
mono/2019-02.  Many of the MSBuild tests were failing because targets
were running when they shouldn't.  Looking at the logs we saw e.g.:

	Input file "/Users/builder/jenkins/workspace/xamarin-android-pr-pipeline-release/xamarin-android/bin/Release/lib/xamarin.android/xbuild-frameworks/MonoAndroid/v1.0/Facades/Microsoft.Win32.Primitives.dll" is newer than output file "obj/Debug/UnnamedProject.dll".

Which meant that the `Microsoft.Win32.Primitives.dll` assembly was
somehow "newer" than a file we just built...  Further investigation
into the mono bundle resulted in an apparent TimeZone issue?

	$ unzip -l android-release-Darwin-cf880be66a2d1cc8ca34e345114a7f420b7b86ea.zip | grep Microsoft.Win32
		4608  04-17-2019 08:54   android-bcl/monodroid/Facades/Microsoft.Win32.Primitives.dll

This file was created on `04-17-2019 08:54`.  However the unit test
build log was built on:

	Build started 4/16/2019 7:50:57 PM.

(*Note*: en-US dates, MM/DD/YYYY.)

Thus the mono bundle appears to have different timestamps.  This is
probably down to something weird on the bots.  In order to try to fix
this we shall `<Touch/>` the files we get from the bundle to make
sure they have the current date, not some (potentially future!) date.