[Xamarin.Android.Build.Tasks] fix warnings from <CompileNativeAssembly/>
Any Xamarin.Android app was generating build output such as:

    _CompileNativeAssemblySources:
      [aarch64-linux-android-as.EXE stderr] typemap.mj.inc: Assembler messages:
      [aarch64-linux-android-as.EXE stderr] typemap.mj.inc:60201: Warning: end of file not at end of a line; newline inserted
      [arm-linux-androideabi-as.EXE stderr] typemap.mj.inc: Assembler messages:
      [arm-linux-androideabi-as.EXE stderr] typemap.mj.inc:60201: Warning: end of file not at end of a line; newline inserted

The fix was we just needed to call `StreamWriter.WriteLine()` once at
the end of the file. I added an `NativeAssemblyDataStream.EndOfFile()`
method to call for this instead of `Flush()`.

I also updated log messages in `<CompileNativeAssembly/>` to be more
precise:

* stdout should probably use `LogDebugMessage` for
  `MessageImportance.Low`.
* stderr should probably use `MessageImportance.High`.

I added a test that checks for MSBuild warnings in general, and that
we aren't getting messages like these anymore.