[build] Adjust dependencies for Ubuntu 19 (#3006)
Ubuntu 19 removed OpenJDK 8 and thus we can no longer depend on distro packages
to install it. OpenJDK 8 is a requirement for Android development but, alas,
there's no replacement in the Ubuntu 19 (or Debian 10+) repositories for this
version. At some point we will start provisioning our own version of OpenJDK 8
but, for now, user needs to install OpenJDK 8 on their own. A recommended
alternative is https://docs.aws.amazon.com/corretto/latest/corretto-8-ug/downloads-list.html

Remove dependency on openjdk-8-jdk and change `libzip4` dependency to
`libzip-dev` in order to install the latest version of libzip.