[MSBuildDeviceIntegration] Fix the Install/InstantRun tests.
We made a change where `SignAndroidPackage` no longer
automatically calls `Build`. This was done to help
speed up the IDE and to save it having to call `Build`
all over again. There was an assumption in the `Install`
tests where we assumed `Build` was being called. As a
result a bunch of tests failed with our favourite
`LinkAssemblies` error.

	error MSB4018: System.IO.FileNotFoundException: Could not load assembly 'UnnamedProject, Version=0.0.0.0, Culture=neutral, PublicKeyToken='. Perhaps it doesn't exist in the Mono for Android profile?

So lets call `Build` before we do any of the Installation
tests.

Next up it seems that Nunit resolves all the test cases
BEFORE the `[OnTimeSetup]` runs. As a result the value
for `CommercialBuildAvailable` was always `false`.
So lets calculate that in a static constructor so that
it will have the correct value when nunit figures out
the test cases.