[tests] Add XF test to _ApkTestProjectBundle
When fixing the AOT measurements
(https://github.com/xamarin/xamarin-android/pull/3017), I didn't
notice we have same issue with the Bundle flavor.

We can either not run XF in Bundle flavor or rebuild it before
run. The XF apps are very common, so I think it makes sense to run the
test and thus we need to add it to the `_ApkTestProjectBundle` item
group so that it is rebuild before running.