[tests] Fix the Install/InstantRun tests. (#3023)
Commit 76fb90e0 changed the `SignAndroidPackage` target so that it
would *not* depend upon the `Build` target, when building within an
IDE (`$(BuildingInsideVisualStudio)`=True).  This was done to help
speed up the IDE and to save it having to call the `Build` target all
over again.

There was an assumption in the `Install` tests where we assumed
`Build` was being called.  As a result, a bunch of tests failed with
our favourite `<LinkAssemblies/>` error.

	error MSB4018: System.IO.FileNotFoundException: Could not load assembly 'UnnamedProject, Version=0.0.0.0, Culture=neutral, PublicKeyToken='. Perhaps it doesn't exist in the Mono for Android profile?

To fix this, explicitly call the `Build` target as part of the
Installation-related tests.

Next up it seems that NUnit resolves all the test cases BEFORE the
`[OnTimeSetup]` runs.  As a result the value for
`BaseTest.SetUp.CommercialBuildAvailable` was always `false`.
Instead, calculate the `CommercialBuildAvailable` value a static
constructor so that it will have the correct value when nunit figures
out the test cases.