[Xamarin.Android.Build.Tasks] first styleable resource id is always incorrect
Fixes #3049

This commit also reworks the id generation of the
ManagedResourceParser to try replicate the
ids which aapt2 produces. As a result if the managed
parser generates the code from the R.txt file or
directly from the resources, the resulting designer.cs
file will be identical most of the time. There does seem
to be an odd ordering issue where the R.txt file does
not seem to order the items by name as the Id's are
sometimes switched.