[Xamarin.Android.Build.Tasks] Add metadata to library imports (#2984)
One of the hard things to do when diagnosing issues is figuring out
where a certain file came from.  Because of the path and command line
limits on Windows we have to use shortened paths by default,
resulting in paths such as:

	obj/Debug/lp/32/res
	obj/Debug/lp/42/classes.jar
	obj/Debug/lp/55/assets

While this is great for path lengths, it does not tell use which
NuGet or Library project those items came from.  There is a map file
which does hold the mapping information,  however it is not a
straightforward process to look at that file and then work out the
path to the assembly or `.aar`.

Improve matters by adding a new piece of metadata to each item
specifying the `%(OriginalFile)` which the item came from.  This will
either be a full path to an Assembly, for items which are extracted 
from an `@(EmbeddedResource)`, or a full path to an `.aar` file,
which was downloaded with a NuGet.

With this information diagnosing issues might get a bit easier.