[Xamarin.Android.Build.Tasks] Support Filtering Assemblies (#2899)
As part of the AndroidX migration system we need to be able to skip
over the older support library assemblies for certain build steps.

Add support for the following metadata on the `@(ReferencePath)`
and `@(ReferenceDependencyPaths)` ItemGroups:

  * `%(AndroidSkipAddToPackage)`: Used by the `<BuildApk/>` task to
    skip over the library and *not* add it to the final `.apk`.

  * `%(AndroidSkipJavaStubGeneration)`: "Boolean" value; when True,
    will prevent Java code from being generated for the library in
    question.

  * `%(AndroidSkipResourceExtraction)`: will prevent the resources in
    that library from being extracted into `$(IntermediateOutputPath)`.

  * `%(AndroidSkipResourceProcessing)`: renamed
    `%(SkipAndroidResourceProcessing)` metadata item (5f27d0ca).
    This is used to skip case conversion on those resources.

To use these new metadata items you can define your own
`<ItemGroup/>` as follows:

	<ItemGroup>
	  <_MyStuff Include="Xamarin.Android.Support.v4">
	    <AndroidSkipAddToPackage>True</AndroidSkipAddToPackage>
	  </_MyStuff>
	</ItemGroup>

Then add this new `<ItemGroup/>` to
`@(AndroidCustomMetaDataForReferences)`:

	<ItemGroup>
	  <AndroidCustomMetaDataForReferences Include="@(_MyStuff)" />
	</ItemGroup>

The `@(AndroidCustomMetaDataForReferences)` ItemGroup will then be
processed by the `_AddAndroidCustomMetaData` target.  This will add
the new metadata values to the `@(ReferencePath)` and
`@(ReferenceDependencyPaths)` ItemGroups.