[tests] fixes for App Bundle tests (#3031)
If run the `BundleToolTests` against an x86 emulator, the tests fail:

	[BT : 0.8.0] error : The app doesn't support ABI architectures of the device. Device ABIs: [x86], app ABIs: [armeabi-v7a, arm64-v8a].

I missed this originally because I was using a device.

In `Release` mode, we don't get x86 by default, so I added:

	var abis = new string [] { "armeabi-v7a", "arm64-v8a", "x86" };
	project.SetProperty (KnownProperties.AndroidSupportedAbis, string.Join (";", abis));

Most of our MSBuild tests that deploy are doing this.  I also added
to our list of assertions to check for x86 `.so` files.

Another problem I found was that this directory was not getting
deleted after successful test runs:

	bin\TestDebug\temp\BundleToolTests

I changed the behavior of this test to default to using
`cleanupOnDispose: true`, and toggle the value to `false` if a test
fails.