[xabuild] Create symlink to MSBuild's `Current` "version" directory (#3027)
Newer MSBuild versions are moving to use a `Current` directory
instead of a versioned one (e.g. `14.0`, `15.0`) to keep its binaries,
targets etc.  Without pointing to that directory in a symlink from
`bin/$(Configuration)/lib/xamarin.android/xbuild`, the build will
fail on Unix with an error similar to:

	bin/Debug/lib/xamarin.android/xbuild/Xamarin/Android/Xamarin.Android.Aapt2.targets(31,3):
	error MSB3191: Unable to create directory "/Debug/". Access to the path '/Debug/' is denied.

Adjust `XABuildPaths.cs` to first look for the `Current` directory
and only then for the old `15.0` one.