[Xamarin.Android.Build.Tasks] don't copy to linksrc for Debug builds
While investigating performance of the `_LinkAssembliesNoShrink`
MSBuild target, I noticed it has the following behavior:

* Take all assemblies from `@(ResolvedAssemblies)`. These are their
  original locations from after `Build`. Some from `bin\Debug`, some
  from `$(UserProfile)\.nuget`, etc.
* Output assemblies directly to be packaged, in:
  `$(IntermediateOutputPath)android\assets`

After a build, however, there was a `linksrc` directory containing all
assemblies. For the Xamarin.Forms app in this repo, it was ~69.7MB and
had 77 files.

But this directory appears to be completely unused when
`$(AndroidLinkMode)` is `None`?

That means we could skip a few MSBuild targets in this case:

* `_CopyIntermediateAssemblies`
* `_CopyPdbFiles`
* `_CopyMdbFiles`

An initial build of the Xamarin.Forms project in this repo used to
take:

     16 ms  _CopyMdbFiles                              1 calls
     77 ms  _CopyPdbFiles                              1 calls
    202 ms  _CopyIntermediateAssemblies                1 calls

So this looks like it saves ~275ms on debug builds. I suspect the
savings would be even better on slower machines (or without SSDs).