[Xamarin.Android.Build.Tasks] Make XA0115 error more actionable (#2967)
Context: https://forums.xamarin.com/discussion/146174/vs2019-error-this-abi-is-no-longer-supported
Context: https://developercommunity.visualstudio.com/comments/485766/view.html

Change the wording of the `armeabi` build error to make it more
actionable at the cost of also making it a bit longer.

In addition to the questions about this error from users on the
forums and Developer Community, another motivation for this change is
that compared to back when I added the original wording for this
message, I have now had more practice updating other error messages
and thinking about how to make them as actionable as possible on
their own.

Unfortunately the correct action to take in the IDE in this case is a
little confusing because the property page in Visual Studio on
Windows is currently different from the property page in
Visual Studio for Mac.  For example, in Visual Studio for Mac, the
**Advanced** tab of the **Android Build** page will show a checkbox
for `armeabi` if that setting is present in the
`$(AndroidSupportedAbis)` property in the `.csproj` file, so users
can just un-check that option to remove the `armeabi` value.  But on
Windows, the **Advanced** modal window under the **Android Options**
page now *never* shows a checkbox for `armeabi`, so the easiest way
to remove the value is to change and change back one of the other
ABIs and save the properties.  That rewrites
`$(AndroidSupportedAbis)` to contain only the values that are *shown*
as selected in the modal window.