[Xamarin.Android.Build.Tasks] Invalid Resources generated for DesignTime
Fixes #3007

Xamarin.Forms introduced some new `android:id` values in the
`Xamarin.Forms.Platform.Android` assembly [1]. These include
a period `.` in the identifier, such as

	android:id="@+id/bottomtab.tabbar"
	android:id="@+id/toolbar.one"

When processing these with `aapt` or `aapt2` these identifiers
are "fixed up" to be valid java identifiers. The same is
NOT true for the `ManagedResourceParser`. As a result
the very first design time build will fail with the following
error.

	obj\Debug\81\designtime\Resource.designer.cs(5048,21,5048,30): error CS0145: A const field requires a value to be provided
	obj\Debug\81\designtime\Resource.designer.cs(5048,30,5048,31): error CS1003: Syntax error, ','
	obj\Debug\81\designtime\Resource.designer.cs(5048,31,5048,38): error CS1002: ; expected

This is because it is generating the following code in the
`Resource.designer.cs`

	public const int toolbar.one = 2131230899;

which is NOT a valid C# identifier.

This commit adds a new `ResourceIdentifier` class which
is responsible for making sure the identifier we generate
is a valid C# identifier. This logic is based on the
C# language specifications [1]. It also does keyword
detection and prepends an `@` if required. Just in case
a user calls an item `base` or `class` or any other
C# keyword.

[1] https://github.com/xamarin/Xamarin.Forms/blob/master/Xamarin.Forms.Platform.Android/Resources/Layout/BottomTabLayout.axml#L13.
[2] https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/language-specification/lexical-structure#identifiers
[3] https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/language-specification/lexical-structure#keywords