Use system mono 6.0.0.
Trying to figure out the build breakage at:
https://jenkins.mono-project.com/job/xamarin-android-pr-pipeline-release/800/

	error MSB4018: The "ResolveLibraryProjectImports" task failed unexpectedly.
	error MSB4018: System.IO.FileNotFoundException: Could not load assembly 'xunit.runner.utility.netstandard20, Version
=0.0.0.0, Culture=neutral, PublicKeyToken='. Perhaps it doesn't exist in the Mono for Android profile?

...and other similarly bizarre errors.

Use system mono 6.0, in the hopes that this is caused by some bizarro
`msbuild` bug which was fixed.  Additionally, *everything else* will
be using mono 6.0, so we best test that we work with it *anyway*.

Additionally, add a debug printf to help track things down, just in
case.