Fix FileNotFoundException
Context: https://jenkins.mono-project.com/job/xamarin-android-pr-pipeline-release/805/
Context: https://xamarinhq.slack.com/archives/C03CEGRUW/p1556849406495900 (plus next hour...)

Things don't build:

	error MSB4018: The "ResolveLibraryProjectImports" task failed unexpectedly.
	error MSB4018: System.IO.FileNotFoundException: Could not load assembly 'Newtonsoft.Json, Version=0.0.0.0, Culture=neutral, PublicKeyToken='. Perhaps it doesn't exist in the Mono for Android profile?

The cause is a comedy of errors:

 1. Because we're not using the mono archive -- it doesn't exist in
    mono/2018-08 -- we have to build our own `netstandard.dll`.

 2. The `netstandard.dll` that we built prior to f5dadad5 referenced
    `Mono.Android.dll`.

 3. The `<FilterAssemblies/>` task (5ec3e3ab) looks for references to
    `Mono.Android.dll`, which `netstandard.dll` now contains.  Ergo,
    `netstandard.dll` is a `MonoAndroid`-profile assembly!

So far so...okay?

Where it went off the rails is that, for some not yet understood
reason, the MSBuild `<ResolveAssemblyReferences/>` task can produce
*weird* output:

	Output Item(s):
	  ReferenceDependencyPaths=
	      .../lib/xamarin.android/xbuild-frameworks/MonoAndroid/v1.0/Facades/netstandard.dll
	          # fair enough
	          ...
	          HintPath=../../../packages/Newtonsoft.Json.11.0.2/lib/netstandard2.0/Newtonsoft.Json.dll
	          # wat?

After `<ResolveAssemblyReferences/>` executes, the
`%(ReferenceDependencyPaths.HintPath)` value for `netstandard.dll`
refers to `Newtonsoft.Json.dll`.

(I have no words.  This makes no sense at all.)

We see this on master as well!  (at least as of 2019-Apr-25 in
`xa-build-status-v9.2.99.191_Darwin-x86_64_master_b255da27-Release.zip`.)

However, on master the `netstandard` reference is "filtered out" by
`<FilterAssemblies/>`, preventing it from ever making it to
`<ResolveLibraryProjectImports/>` (which does look at `%(HintPath)`,
which is why it's trying to load `Newtonsoft.Json.dll` in the first
place -- because it's trying to load `netstandard.dll`!).

Here, no such filtering occurs -- `netstandard.dll` is a
MonoAndroid-profile assembly! -- so it gets processed, causing
`<ResolveLibraryProjectImports/>` to try to read the file referenced
by `%(HintPath)`, which fails because while the source path exists,
`DirectoryAssemblyResolver` removes that path information to construct
an `AssemblyNameReference` to load from the `SearchDirectories`, none
of which contain `Newtonsoft.Json.dll`.

Now that we know "why" it's failing, how do we fix it?

Let's go with quick-and-stupid: wrap the
`DirectoryAssemblyResolver.GetAssembly()` call around a `try/catch`
and ignore the assembly if it can't be retrieved.

TODO: should we instead/also revert f5dadad5?