Rationale for d16-2-with-mono-2018-10-via-6e8a50ea
(This is an empty commit.)

Context: https://github.com/mono/mono/pull/14307

The [Xamarin.Android Designer][0] is an interesting beast: it runs a
Xamarin.Android-like Java stack on desktop macOS and Windows.  A Java
process is created, and that Java process loads:

  * *Desktop* Mono via `libmonosgen-2.0.dylib` (macOS)` or
    `libmonosgen-2.0.dll` (Windows).`
  * The Xamarin.Android runtime via `libmono-android.debug.dylib`
    (macOS) or `libmono-android.debug.dll` (Windows)
  * Other Java code (to render items to the screen)
  * `Mono.Android.dll` and other user assemblies (to render items to
    the screen).

Being "interesting" means it introduces all kinds of "fun":
`libmono-android.debug.dll` needs to be built *for Windows* (which is
why there's `#if WINDOWS` within `src/monodroid`),
`libmonosgen-2.0.dll` needs to be built for Windows (not a huge
issue), and -- "equally" important but *more* important regarding the
next few paragraphs! -- `mscorlib.dll` needs to target Windows.

`mscorlib.dll` is where things get annoying.  Once Upon A Time™,
Mono's `mscorlib.dll` was platform agnostic: build it for one
platform, it would run everywhere, as it would internally perform
OS-checks to behave appropriate on Unix/Windows/etc.

That hasn't been true for awhile, but the Xamarin.Android build and
install environment still assumes it to be true.  Consequently, a
"Unix" `mscorlib.dll` is used by the Designer on Windows.  This
[*apparently* (?)][1] hasn't blown up catastrophically before, but
with the mono/2019-02 migration (b8c30f9e) we have now entered
"catastrophic" territory: [Designer integration tests on Windows][2]
may fail because `System.Native` cannot be found:

  System.TypeInitializationException: The type initializer for 'System.Console' threw an exception. ---> System.DllNotFoundException: System.Native

Given current release timeframes, d16-2 Preview 1 cannot ship using
mono/2019-02, not without a new [`monodroid_designer` profile][3].

If we can't use mono/2019-02, there are only two alternatives:

 1. Use mono/2018-10, from *before* commit b8c30f9e, or
 2. Use mono/2018-08, as is used in [d16-1][4].

The problem with mono/2018-10 is that while a Unix-targeting
`mscorlib.dll` appears to work properly on Windows, parts of
`System.dll` won't.  In particular, `SafeFileHandle` uses
`System.Native`, and use of it would result in
`DllNotFoundException`s.  (This is *presumably* unlikely to happen,
granted, but the fact that it *could* happen is worrying.)

This makes mono/2018-08 look like a good idea, but that is in fact
[very painful][5], resulting in [862 unit test failures][6].

Time is short.  We're using mono/2018-10, as that
[appears to work][7].

*Note*: the d16-2-with-mono-2018-10-via-6e8a50ea branch is branched
from xamarin/xamarin-android/d16-2@6e8a50ea.

[0]: https://docs.microsoft.com/en-us/xamarin/android/user-interface/android-designer/
[1]: https://developercommunity.visualstudio.com/content/problem/440975/xamarin-xaml-designer-still-fails.html
[2]: https://devdiv.visualstudio.com/DevDiv/_build/results?buildId=2634805
[3]: https://github.com/mono/mono/pull/14307
[4]: https://github.com/xamarin/xamarin-android/commits/d16-1/
[5]: https://github.com/xamarin/xamarin-android/pull/3056
[6]: https://jenkins.mono-project.com/job/xamarin-android-pr-pipeline-release/811/testReport/
[7]: https://github.com/xamarin/xamarin-android/pull/3064